renaming folder `substrate` to `substrate-lib`
