+++ src/Events.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { Feed, Grid } from "semantic-ui-react";
 
-import { useSubstrate } from "./substrate";
+import { useSubstrate } from "./substrate-lib";
 
 export default function Events(props) {
   const { api } = useSubstrate();
