+++ src/Metadata.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { Grid, Modal, Button, Card } from "semantic-ui-react";
 
-import { useSubstrate } from "./substrate";
+import { useSubstrate } from "./substrate-lib";
 
 export default function Metadata(props) {
   const { api } = useSubstrate();
