+++ src/NodeInfo.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { Card, Icon, Grid } from "semantic-ui-react";
 
-import { useSubstrate } from "./substrate";
+import { useSubstrate } from "./substrate-lib";
 
 export default function NodeInfo(props) {
   const { api } = useSubstrate();
