+++ src/config/production.json
@@ -1,3 +1,3 @@
 {
-  "PROVIDER_SOCKET": "wss://dev-node.substrate.dev:9944"
+  "PROVIDER_SOCKET": ""
 }
